Add REARCENTER, for distinguising better between 6.1 and 7.1 channels. Make KMix honour all possible channel types instead of only FRONTLEFT and FRONTRIGHT to the ALSA backend (maps to 9 different SND_MIXER_SCHN_* values). CCBUGS: 180378
svn path=/trunk/KDE/kdemultimedia/kmix/; revision=913108

+++ mixer.cpp
@@ -421,7 +421,7 @@ MixDevice* Mixer::getMixdeviceById( const QString& mixdeviceID )
 {
    MixDevice* md = 0;
    int num = _mixerBackend->id2num(mixdeviceID);
-   if ( num!=-1 && num < size() ) {
+   if ( num!=-1 && num < (int)size() ) {
       md = (*this)[num];
    }
    return md;
